refactoring merkle code
