+++ package.json
@@ -20,7 +20,7 @@
     "lintall:fix": "yarn lint:fix && yarn pretty:fix && yarn sol:fix",
     "format": "yarn lintall:fix",
     "merkle": "npx ts-node scripts/generateMerkle.ts",
-    "unlcaimed": "npx ts-node scripts/unclaimed.ts"
+    "unclaimed": "npx ts-node scripts/unclaimed.ts"
   },
   "repository": {
     "type": "git",
